proxy/requirements.txt
requests==2.31.0
flask==2.3.3
python-dotenv
